ai-speech-coach/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.2
sqlalchemy>=2.0.21
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
nltk>=3.8.1
scikit-learn>=1.3.1
pandas>=2.1.1
numpy>=1.26.0
mcp>=1.2.0
pytest>=7.4.2